+++ runtime/compiler/z/codegen/J9CodeGenerator.cpp
@@ -41,7 +41,8 @@
 #include "il/Node.hpp"
 #include "il/Node_inlines.hpp"
 #include "z/codegen/J9S390PrivateLinkage.hpp"
-#include "z/codegen/J9S390SystemLinkage.hpp"
+#include "z/codegen/J9SystemLinkageLinux.hpp"
+#include "z/codegen/J9SystemLinkagezOS.hpp"
 #include "z/codegen/J9S390CHelperLinkage.hpp"
 #include "z/codegen/S390GenerateInstructions.hpp"
 #include "z/codegen/S390Recompilation.hpp"
